requirements.txt
pip==25.0.1
setuptools==59.5.0
gunicorn==23.0.0
Flask==2.3.3
mysql-connector==2.2.9
matplotlib==3.8.3
numpy==1.22.0
opencv-python==4.9.0.80
pillow==10.2.0
python-dotenv==1.0.1
PyMySQL==1.1.1
requests==2.31.0
scikit-learn==1.6.1
scipy==1.11.4